// commission/src/modules/grpc/proto/commission.pb.ts
export interface CreateCommissionSummaryRequest {
  commission: CommissionFilters | undefined;
  product: Product | undefined;
  calculationSettings: CalculationSettings | undefined;
  promoCode: PromoCode | undefined;
  order: Order | undefined;
  paymentModuleName: string;
  paymentOption: PaymentOption | undefined;
  reservation: Reservation | undefined;
  addOns: AddOn[];
  financingRequest: FinancingRequest | undefined;
}

export interface UpdateSellerCommissionRequest {
  product: Product | undefined;
  sellerNewCommission: number;
}

export interface ForceUpdateCommissionRequest {
  productId: string;
  orderId: string;
  grandTotal: number;
  payout: number;
  discount: number;
  buyerCommission: number;
  sellPrice: number;
}

export interface UpdateSellPriceRequest {
  product: Product | undefined;
}

export interface MigrateCommissionSummaryRequest {
  commission: CommissionFilters | undefined;
  product: Product | undefined;
  calculationSettings: MigrationCalculationSettings | undefined;
  promoCode: PromoCode | undefined;
  order: Order | undefined;
  paymentModuleName: string;
}

export interface Order {
  id: string;
}

export interface AddOn {
  id: string;
  addOnPrice: number;
}

export interface PaymentOption {
  id: string;
  paymentProvider: string;
  paymentCardType: string;
}

export interface CommissionFilters {
  userType: string;
  isBuyer: boolean;
}

export interface Product {
  id: string;
  sellPrice: number;
  priceRange: string;
  source: string;
  categoryId: string;
  modelId: string;
  varientId: string;
  conditionId: string;
}

export interface CalculationSettings {
  vatPercentage: number;
  applyDeliveryFeeSPPs: boolean;
  applyDeliveryFeeMPPs: boolean;
  applyDeliveryFee: boolean;
  deliveryFeeThreshold: number;
  deliveryFee: number;
  referralFixedAmount: number;
  applyReservation: boolean;
  applyFinancing: boolean;
}

export interface MigrationCalculationSettings {
  vatPercentage: number;
  applyDeliveryFeeSPPs: boolean;
  applyDeliveryFeeMPPs: boolean;
  applyDeliveryFee: boolean;
  deliveryFeeThreshold: number;
  deliveryFee: number;
  referralFixedAmount: number;
  buyerCommissionPercentage: number;
  sellerCommissionPercentage: number;
  priceQualityExtraCommission: number;
}

export interface Reservation {
  reservationAmount: number;
  reservationRemainingAmount: number;
}

export interface FinancingRequest {
  amount: number;
}

export interface PromoCode {
  promoLimit: number;
  type: string;
  generator: string;
  discount: number;
  percentage: number;
}
// ...
export interface CalculateAddonSummaryRequest {
  productPrice: number;
  addonSummaryCalculateData: AddonSummaryCalculateData[];
}

export interface AddonSummaryCalculateData {
  priceType: string;
  addonPrice: number;
}

export interface CalculateAddonSummaryResponse {
  addOnsTotal: number;
  addOnsVat: number;
  addOnsGrandTotal: number;
}